+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/internal/tls/SslContextBuildException.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.proxy.config.tls;
+package io.kroxylicious.proxy.internal.tls;
 
 public class SslContextBuildException extends RuntimeException {
