+++ kroxylicious-runtime/src/test/resources/io/kroxylicious/proxy/internal/tls/client.jks
