+++ kroxylicious-runtime/src/test/resources/io/kroxylicious/proxy/internal/tls/server_crt_encrypted_key.pem
