+++ kroxylicious-runtime/src/test/resources/io/kroxylicious/proxy/internal/tls/server_encrypted.key
